Improve the way to handle the settings by storing to sessionStorage
Now, the extension try to load settings from sessionStorage when it starts. And it saves when you change settings. So the setting values don't go away even if you enter some URL and browse then come back the board.

Popup shows the extension is alive or not on current page.